+++ internal/provider/resource_sdwan_cisco_bgp_feature_template_test.go
@@ -65,11 +65,7 @@ func TestAccSdwanCiscoBGPFeatureTemplate(t *testing.T) {
 					resource.TestCheckResourceAttr("sdwan_cisco_bgp_feature_template.test", "address_families.0.ipv4_aggregate_addresses.0.prefix", "10.0.0.0/8"),
 					resource.TestCheckResourceAttr("sdwan_cisco_bgp_feature_template.test", "address_families.0.ipv4_aggregate_addresses.0.as_set_path", "true"),
 					resource.TestCheckResourceAttr("sdwan_cisco_bgp_feature_template.test", "address_families.0.ipv4_aggregate_addresses.0.summary_only", "true"),
-					resource.TestCheckResourceAttr("sdwan_cisco_bgp_feature_template.test", "address_families.0.ipv6_aggregate_addresses.0.prefix", "10.0.0.0/8"),
-					resource.TestCheckResourceAttr("sdwan_cisco_bgp_feature_template.test", "address_families.0.ipv6_aggregate_addresses.0.as_set_path", "true"),
-					resource.TestCheckResourceAttr("sdwan_cisco_bgp_feature_template.test", "address_families.0.ipv6_aggregate_addresses.0.summary_only", "true"),
 					resource.TestCheckResourceAttr("sdwan_cisco_bgp_feature_template.test", "address_families.0.ipv4_networks.0.prefix", "10.2.2.0/24"),
-					resource.TestCheckResourceAttr("sdwan_cisco_bgp_feature_template.test", "address_families.0.ipv6_networks.0.prefix", "10.2.2.0/24"),
 					resource.TestCheckResourceAttr("sdwan_cisco_bgp_feature_template.test", "address_families.0.maximum_paths", "8"),
 					resource.TestCheckResourceAttr("sdwan_cisco_bgp_feature_template.test", "address_families.0.default_information_originate", "true"),
 					resource.TestCheckResourceAttr("sdwan_cisco_bgp_feature_template.test", "address_families.0.table_map_policy", "MAP1"),
@@ -188,15 +184,7 @@ func testAccSdwanCiscoBGPFeatureTemplateConfig_all() string {
 				as_set_path = true
 				summary_only = true
 			}]
-			ipv6_aggregate_addresses = [{
-				prefix = "10.0.0.0/8"
-				as_set_path = true
-				summary_only = true
-			}]
 			ipv4_networks = [{
-				prefix = "10.2.2.0/24"
-			}]
-			ipv6_networks = [{
 				prefix = "10.2.2.0/24"
 			}]
 			maximum_paths = 8